EWD-22647 - Learning object -> Learning content EWD-22638 - Delete non-existing experience
Former-commit-id: e0ef527e509db2fb4814ef565c56fe6b0001bf62

+++ sources/easygenerator.Web/Controllers/Api/ExperienceController.cs
@@ -46,6 +46,12 @@ namespace easygenerator.Web.Controllers.Api
         [HttpPost]
         public ActionResult Delete(Experience experience)
         {
+            if (experience == null)
+            {
+                return JsonLocalizableError(Constants.Errors.ExperienceNotFoundError,
+                                            Constants.Errors.ExperienceNotFoundResourceKey);
+            }
+
             _repository.Remove(experience);
 
             return JsonSuccess();
